+++ dejafu/Test/DejaFu/Schedule.hs
@@ -11,12 +11,6 @@ module Test.DejaFu.Schedule
   ( -- * Scheduling
     Scheduler(..)
 
-  , Decision(..)
-  , tidOf
-  , decisionOf
-
-  , NonEmpty(..)
-
   -- ** Preemptive
   , randomSched
   , roundRobinSched
@@ -56,37 +50,6 @@ newtype Scheduler state = Scheduler
     -> state
     -> (Maybe ThreadId, state)
   }
-
--------------------------------------------------------------------------------
--- Scheduling decisions
-
--- | Get the resultant thread identifier of a 'Decision', with a default case
--- for 'Continue'.
---
--- @since 0.5.0.0
-tidOf :: ThreadId -> Decision -> ThreadId
-tidOf _ (Start t)    = t
-tidOf _ (SwitchTo t) = t
-tidOf tid _          = tid
-
--- | Get the 'Decision' that would have resulted in this thread
--- identifier, given a prior thread (if any) and collection of threads
--- which are unblocked at this point.
---
--- @since 0.5.0.0
-decisionOf :: Foldable f
-  => Maybe ThreadId
-  -- ^ The prior thread.
-  -> f ThreadId
-  -- ^ The threads.
-  -> ThreadId
-  -- ^ The current thread.
-  -> Decision
-decisionOf Nothing _ chosen = Start chosen
-decisionOf (Just prior) runnable chosen
-  | prior == chosen = Continue
-  | prior `elem` runnable = SwitchTo chosen
-  | otherwise = Start chosen
 
 -------------------------------------------------------------------------------
 -- Preemptive